update: split post_message
